+++ packages/instant-apps-components/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0-beta.297
+
+### instant-apps-scoreboard
+
+- Add property to determine whether stats query is queried against feature layer service or feature layer view
+
 ## v1.0.0-beta.296
 
 ### instant-apps-export